Integrated i18n-abide. Removed lib/i18n.js. Fixes #1322.
* Fixes #2644 and Bug# 806649
* Removes eval usage
* Fixes format API drift %(foo) versus %(foo)s
