+++ bin/static
@@ -19,7 +19,8 @@ connect_fonts_opensans = require('connect-fonts-opensans');
 const
 assets = require('../lib/static_resources').all,
 cachify = require('connect-cachify'),
-i18n = require('../lib/i18n.js'),
+i18n = require('i18n-abide'),
+i18n_check = require('../lib/i18n_client_check'),
 wsapi = require('../lib/wsapi.js'),
 httputils = require('../lib/httputils.js'),
 db = require('../lib/db.js'),
@@ -66,6 +67,8 @@ app.use(i18n.abide({
   translation_directory: config.get('translation_directory'),
   disable_locale_check: config.get('disable_locale_check')
 }));
+
+i18n_check();
 
 var statsd_config = config.get('statsd');
 if (statsd_config && statsd_config.enabled) {
