+++ docs/I18N.md
@@ -52,22 +52,15 @@ config file which you'll like to validate. Examples:
 
 ## Debugging
 
-If code is evaluated in node.js (server-side) then node-gettext is providing the string
-translation. Strings are from the `messages.mo` files under the locale directory. MO files
-are binary, compiled from the PO files.
+If code is evaluated in node.js (server-side) then translation is provided from the
+`i18n/{locale}/messages.json` file which contains the translation. These JSON files come
+from PO files.
 
-If code is evaluated on the client-side, then `resources/static/shared/gettext.js` is in
-the house. Strings are from `resources/static/i18n` JSON files.
+If code is evaluated on the client-side, then `static/gettext.js` is in
+the house... strings are `i18n/{locale}/client.json` files.
 
 If code is evaluated in your head, then clearly we are post-singularity. Why are you
 still using gettext?
-
-Use the `eo` locale for development and debugging. It is auto-translated with:
-
-    for catalog in messages client; do
-      echo "Translating ${catalog}.po"
-      podebug --rewrite=flipped -i locale/templates/LC_MESSAGES/${catalog}.pot -o locale/eo/LC_MESSAGES/${catalog}.po
-    done
 
 ## New Locales
