+++ lib/static_resources.js
@@ -1,4 +1,4 @@
-var i18n = require('./i18n'),
+var i18n = require('i18n-abide'),
     und = require('underscore');
 
 /**
