+++ lib/wsapi.js
@@ -28,7 +28,8 @@ path = require('path'),
 validate = require('./validate'),
 statsd = require('./statsd'),
 bcrypt = require('./bcrypt'),
-i18n = require('./i18n'),
+i18n = require('i18n-abide'),
+i18n_check = require('./i18n_client_check'),
 db = require('./db'),
 http = require('http'),
 https = require('https');
@@ -39,6 +40,8 @@ var abide = i18n.abide({
   translation_directory: config.get('translation_directory'),
   disable_locale_check: config.get('disable_locale_check')
 });
+
+i18n_check();
 
 const COOKIE_SECRET = secrets.hydrateSecret('browserid_cookie', config.get('var_path'));
 var COOKIE_KEY = 'browserid_state';
