+++ resources/static/common/js/gettext.js
@@ -34,7 +34,7 @@
           if (_.isArray(obj) || named === false) {
             return fmt.replace(/%s/g, function(match){return String(obj.shift());});
           } else if (_.isObject(obj) || named === true) {
-            return fmt.replace(/%\(\s*([^)]+)\s*\)/g, function(m, v){
+            return fmt.replace(/%\(\s*([^)]+)\s*\)s/g, function(m, v){
               return String(obj[v]);
             });
           }
