+++ resources/static/dialog/views/primary_offline.ejs
@@ -4,7 +4,7 @@
 
   <div id="primary_offline">
     <p>
-      <%= format(gettext('%(primary) is not responding. Please wait a few minutes and try again.'), { primary: idpName }) %>
+      <%= format(gettext('%(primary)s is not responding. Please wait a few minutes and try again.'), { primary: idpName }) %>
     </p>
     <button id="primary_offline_confirm"><%= gettext('Cancel') %></button>
   </div>
