+++ resources/static/dialog/views/rp_info.ejs
@@ -19,7 +19,7 @@
 
 <% if(privacyPolicy && termsOfService) { %>
   <p id="rptospp" class="tospp">
-    <%- format(gettext("By proceeding, you agree to %(site)'s <a %(terms)>Terms</a> and <a %(privacy)>Privacy Policy</a>."),
+    <%- format(gettext("By proceeding, you agree to %(site)s's <a %(terms)s>Terms</a> and <a %(privacy)s>Privacy Policy</a>."),
                {
                  terms: 'href="' + termsOfService + '" id="rp_tos" target="_blank"',
                  privacy: 'href="' + privacyPolicy + '" id="rp_pp" target="_blank"',
