+++ resources/static/dialog/views/set_password.ejs
@@ -11,7 +11,7 @@
           <li>
             <% if (!password_reset) { %>
               <% if (transition_no_password) { %>
-                <%- format(gettext("%(idp) no longer allows you to log into Persona with your %(idp) password. Please create a new password to use with your Persona account."), {
+                <%- format(gettext("%(idp)s no longer allows you to log into Persona with your %(idp)s password. Please create a new password to use with your Persona account."), {
                   idp: "<strong>" + escape(domain) + "</strong>"
                 }) %>
               <% } else { %>
