+++ resources/static/dialog/views/upgrade_to_primary_user.ejs
@@ -2,6 +2,6 @@
       License, v. 2.0. If a copy of the MPL was not distributed with this
       file, You can obtain one at http://mozilla.org/MPL/2.0/. */ %>
   <div id="upgrade_to_primary">
-    <p><%= format(gettext('Your %(emailProvider) account has been upgraded to work even better with Persona. We will redirect you to %(emailProvider). Once you verify your account there, you will be signed into %(aWebsite).'), { emailProvider: idpName, aWebsite: siteName }) %></p>
+    <p><%= format(gettext('Your %(emailProvider)s account has been upgraded to work even better with Persona. We will redirect you to %(emailProvider)s. Once you verify your account there, you will be signed into %(aWebsite)s.'), { emailProvider: idpName, aWebsite: siteName }) %></p>
     <p><button id="upgradeToPrimary" class="upgrade_to"><%= format(gettext("sign in with %s"), [idpName]) %></button></p>
   </div>
