+++ resources/views/about.ejs
@@ -9,7 +9,7 @@
             <article class="blurb">
                 <div class="info first">
                     <h1><%- gettext('Persona replaces multiple passwords') %></h1>
-                    <p><%- format(gettext('Sites such as <a %(timesLink)>%(timesName)</a>, <a %(openphotoLink)>%(openphotoName)</a> and <a %(voostLink)>%(voostName)</a> use Persona instead of usernames to sign you in.'),
+                    <p><%- format(gettext('Sites such as <a %(timesLink)s>%(timesName)s</a>, <a %(openphotoLink)s>%(openphotoName)s</a> and <a %(voostLink)s>%(voostName)s</a> use Persona instead of usernames to sign you in.'),
                                   {
                                     timesLink: 'href="http://crossword.thetimes.co.uk/" target="_blank"',
                                     timesName: 'The Times Crossword',
