+++ resources/views/forgot.ejs
@@ -12,7 +12,7 @@
                   <h2><%- gettext('Confirm your email address') %></h2>
 
                   <p>
-                    <%- format(gettext('Check your email at <strong %(checkId)></strong>.'), { checkId: 'id="sentToEmail"' }) %>
+                    <%- format(gettext('Check your email at <strong %(checkId)s></strong>.'), { checkId: 'id="sentToEmail"' }) %>
                   </p>
 
                   <p>
