+++ resources/views/signin.ejs
@@ -13,7 +13,7 @@
                   <h2><%- gettext('Confirm your email address') %></h2>
 
                   <p>
-                    <%- format(gettext('Check your email at <strong %(emailId)></strong>.'), { emailId: 'id="sentToEmail"' }) %>
+                    <%- format(gettext('Check your email at <strong %(emailId)s></strong>.'), { emailId: 'id="sentToEmail"' }) %>
                   </p>
 
                   <p>
@@ -81,7 +81,7 @@
                  signed in before they enter their email address. -->
             <p class="submit tospp cf">
                <%- format(
-                    gettext('By proceeding, you agree to %(persona)\'s <a %(termsLink)>Terms</a> and <a %(privacyLink)>Privacy Policy</a>.'),
+                    gettext('By proceeding, you agree to %(persona)s\'s <a %(termsLink)s>Terms</a> and <a %(privacyLink)s>Privacy Policy</a>.'),
                          { persona: "Persona",
                            termsLink: 'href="https://login.persona.org/tos" target="_new"',
                            privacyLink: 'href="https://login.persona.org/privacy" target="_new"',
@@ -105,7 +105,7 @@
             </div>
 
             <div class="submit cf forminputs verify_primary">
-              <%- format(gettext('To verify that you own <strong %(emailId)>this address</strong>, you must sign in with your provider. A new window will be opened.'),
+              <%- format(gettext('To verify that you own <strong %(emailId)s>this address</strong>, you must sign in with your provider. A new window will be opened.'),
                          { emailId: 'id="primary_email"' }) %>
 
               <p class="submit">
