+++ resources/views/unsupported_dialog.ejs
@@ -13,11 +13,11 @@
       </a>
 
       <p>
-        <%- format(gettext('Persona works with <a %(getFirefoxLink)>Firefox</a>'), { getFirefoxLink: 'href="http://getfirefox.com" target="_blank" title="Get Firefox"' }) %>
+        <%- format(gettext('Persona works with <a %(getFirefoxLink)s>Firefox</a>'), { getFirefoxLink: 'href="http://getfirefox.com" target="_blank" title="Get Firefox"' }) %>
       </p>
 
       <p class="lighter">
-       <%- format(gettext('and other <a %(otherBrowserLink)>modern browsers.</a>'), { otherBrowserLink: 'href="http://whatbrowser.org" target="_blank"' }) %>
+       <%- format(gettext('and other <a %(otherBrowserLink)s>modern browsers.</a>'), { otherBrowserLink: 'href="http://whatbrowser.org" target="_blank"' }) %>
       </p>
 
   </section>
