+++ scripts/browserid.spec
@@ -23,7 +23,6 @@ persona server & web home for persona.org
 %build
 npm install
 export PATH=$PWD/node_modules/.bin:$PATH
-./locale/compile-mo.sh locale/
 ./locale/compile-json.sh locale/ resources/static/i18n/
 env CONFIG_FILES=$PWD/config/l10n-all.json scripts/compress
 rm -r resources/static/build resources/static/test
