+++ tests/i18n-tests.js
@@ -8,7 +8,7 @@ require('./lib/test_env.js');
 
 const assert = require('assert'),
       vows = require('vows'),
-      i18n = require('../lib/i18n'),
+      i18n = require('i18n-abide'),
       start_stop = require('./lib/start-stop.js'),
       wsapi = require('./lib/wsapi.js'),
       http = require('http'),
@@ -17,90 +17,6 @@ const assert = require('assert'),
 var suite = vows.describe('i18n');
 
 suite.options.error = false;
-
-suite.addBatch({
-  "format a string with place values": {
-    topic: function () {
-      return i18n.format("%s %s!", ["Hello", "World"]);
-    },
-    "was interpolated": function (str) {
-      assert.equal(str, "Hello World!");
-    }
-  }
-});
-
-suite.addBatch({
-  "format a string with named values": {
-    topic: function () {
-      var params = { salutation: "Hello", place: "World" };
-      return i18n.format("%(salutation) %(place)!", params);
-    },
-    "was interpolated": function (str) {
-      assert.equal(str, "Hello World!");
-    }
-  }
-});
-
-suite.addBatch({
-  "format a string without interpolation": {
-    topic: function () {
-      return i18n.format("Hello World!");
-    },
-    "was interpolated": function (str) {
-      assert.equal(str, "Hello World!");
-    }
-  },
-  "format a null": {
-    topic: function () {
-      return i18n.format(null);
-    },
-    "was interpolated": function (str) {
-      assert.equal(str, "");
-    }
-  }
-});
-
-suite.addBatch({
-  "We find exact language match": {
-    topic: function () {
-      var accept = 'pa,sv;q=0.8,fi;q=0.7,it-ch;q=0.5,en-us;q=0.3,en;q=0.2';
-      var supported = ['af', 'en-US', 'pa'];
-      var def = 'en-US';
-      return i18n.bestLanguage(
-          i18n.parseAcceptLanguage(accept),
-          supported, def);
-    },
-    "For Punjabi": function (locale) {
-      assert.equal(locale, "pa");
-    }
-  },
-  "Issue#1128 We find best locale even if region doesn't match": {
-    topic: function () {
-      var accept = 'pa-it,sv;q=0.8,fi;q=0.7,it-ch;q=0.5,en-us;q=0.3,en;q=0.2';
-      var supported = ['af', 'en-US', 'pa'];
-      var def = 'en-US';
-      return i18n.bestLanguage(
-          i18n.parseAcceptLanguage(accept),
-          supported, def);
-    },
-    "For Punjabi (India) serve Punjabi": function (locale) {
-      assert.equal(locale, "pa");
-    }
-  },
-  "We don't extend into a region, unless we have an exact match": {
-    topic: function () {
-      var accept = 'pa,sv;q=0.8,fi;q=0.7,it-ch;q=0.5,en-us;q=0.3,en;q=0.2';
-      var supported = ['af', 'en-US', 'pa-IT'];
-      var def = 'en-US';
-      return i18n.bestLanguage(
-          i18n.parseAcceptLanguage(accept),
-          supported, def);
-    },
-    "Don't choose Punjabi (India)": function (locale) {
-      assert.equal(locale, "en-us");
-    }
-  }
-});
 
 // point to test translation files
 process.env['TRANSLATION_DIR'] = path.join(__dirname, "i18n_test_files");
