+++ tests/i18n_test_files/bg/messages.json
@@ -1,5 +1,5 @@
-var json_locale_data = {
-  messages: {
+;var json_locale_data = {
+  "messages": {
     "This is a translation <strong>test</strong> string.": [
       null,
       "Прова?  Прова?  Четери, пет, шещ?"
@@ -9,4 +9,5 @@ var json_locale_data = {
       ""
     ]
   }
-};
+}
+;
